adding new movie genre component, moving app.js back out of components, and updating app.css
