+++ src/App.css
@@ -2,24 +2,18 @@
   border-radius: 6px 0 0 6px;
 }
 
-.movie-genre {
-  transform-origin: top;
-  text-transform: uppercase;
-  position: relative;
-}
-
 .movie-poster {
   max-width: 68px;
   object-fit: cover;
 }
 
+.movie-poster-container {
+  width: 100px;
+}
+
 .movie-poster,
 .movie-poster-container {
   height: 100px;
-}
-
-.movie-poster-container {
-  width: 100px;
 }
 
 .ratings-loader {
