+++ src/index.js
@@ -1,4 +1,4 @@
-import App from './components/App';
+import App from './App';
 import CssBaseline from '@mui/material/CssBaseline';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
